+++ test/helpers.spec.ts
@@ -1,18 +1,35 @@
-import {
-	directoryExists,
-	transformDirSizeThreshold,
-	transformSegmentTime,
-} from '../src/helpers';
+import { mocked } from 'ts-jest/utils';
+import directoryExists from '../src/helpers/directoryExists';
+import dirSize from '../src/helpers/space';
+import fs from 'fs';
+import transformDirSizeThreshold from '../src/helpers/sizeThreshold';
+import transformSegmentTime from '../src/helpers/segmentTime';
+
+jest.mock('fs');
+jest.mock('path');
 
 describe('directoryExists', () => {
-	test('Valid value', () => {
-		expect(directoryExists(__dirname)).toBeTruthy();
-		expect(directoryExists('./invalid-directory')).toBeFalsy();
+	test('exists', () => {
+		// @ts-ignore
+		mocked(fs).lstatSync.mockReturnValue({isDirectory: () => true});
+
+		expect(directoryExists('path')).toBeTruthy();
 	});
 
-	test('Invalid values', () => {
-		expect(() => directoryExists(__filename))
-			.toThrowError(`${__filename} exists but it is not a directory.`);
+	test('does not exist', () => {
+		// @ts-ignore
+		mocked(fs).lstatSync.mockImplementation(() => {
+			throw new Error('no such file or directory');
+		});
+
+		expect(directoryExists('path')).toBeFalsy();
+	});
+
+	test('not a directory', () => {
+		// @ts-ignore
+		mocked(fs).lstatSync.mockReturnValue({isDirectory: () => false});
+
+		expect(() => directoryExists('path')).toThrowError('path exists but it is not a directory.');
 	});
 });
 
@@ -24,4 +41,14 @@ test('transformDirSizeThreshold', () => {
 test('transformSegmentTime', () => {
 	expect(transformSegmentTime(1)).toEqual(1);
 	expect(transformSegmentTime('1m')).toEqual(60);
+});
+
+test('should return directory size in bytes', () => {
+	mocked(fs).readdirSync.mockReturnValue(new Array(3).fill(0));
+	// @ts-ignore
+	mocked(fs).statSync.mockReturnValue({isDirectory: () => false, size: 3});
+
+	const size = dirSize('');
+
+	expect(size).toEqual(9);
 });